Fix binary search bounds in nextGreatestLetter

+++ 744-find-smallest-letter-greater-than-target/solution.cpp
@@ -5,7 +5,7 @@ using namespace std;
 class Solution {
 public:
     char nextGreatestLetter(vector<char>& letters, char target) {
-        int l = 0, h = static_cast<int>(letters.size()) - 1;
+        int l = 0, h = letters.size() - 1;
         char v = letters[0];
         while (l <= h) {
             int m = l + (h - l) / 2;
